Remove project service from game-tools
We don't need to create projects in game-tools like we use to. This
remove the service but keeps the front-end, allowing AGDE to still
act like it can create projects.

In the future, if there is a need to create/open projects a new project
service can be created.

Before all the services have moved away from using project id, it should
be possible for clients to skip over the "create project" call.

Change-Id: Icf8c817f2462ef5816fd1dfd9414f33601678f24