Autoupdate of 2 submodules
*tools/analytics-library updated to 91ead0dd3a197973df0ca57ad8c4e37f1834f838
Add ENABLE_EMULATOR_CONTROL to analytics_enums.proto

Bug: 276981739
Test: existing
Change-Id: I1339709f9f22e701e64be154809c5d73eaf85fcb

*tools/base updated to bd1add767bf09b384bec613519b786f1ae4ad254
Hide emulator control behind a proeprty flag

 Add new BooleanOption ENABLE_EMULATOR_CONTROL, the property:

android.experimental.enableEmulatorControl=true

needs to be present for the feature to be enabled. This indicates that
a user understand it is an experimental feature.

* EmulatorControlConfig is enabled only if both emulatorControl.enable
and ENABLE_EMULATOR_CONTROL are true
* Add new test cases for EmulatorControlConfig

Test: Additional unit tests
Test:
Starting a Gradle Daemon, 4 incompatible and 2 stopped Daemons could not
be reused, use --status for details

FAILURE: Build failed with an exception.

* What went wrong:
Could not determine the dependencies of task
':app:connectedAndroidTest'.
> Could not create task ':app:connectedDebugAndroidTest'.
   > EmulatorControl is an experimental feature and it is disabled by
   default. Please add
   android.experimental.androidTest.enableEmulatorControl=true in your
   gradle.properties to opt-in to this feature.

Bug: 276981739
Change-Id: I8c17c12fe71329f48c449194d62293d2da416cab