Add "@interface" as explicit token for Proguard
This fixes highlighting for a rule like:

```
-keep,allowobfuscation @interface androidx.annotation.Keep
```

Previously, the "@" was marked as invalid.

The best documentation for these rules is at:
https://www.guardsquare.com/manual/configuration/usage#classspecification

That rule specifies that the class type can be specified as "interface",
"class", or "enum". But when running proguard and there's an invalid
value there, the error message indicates that "@interface" is also a
valid value.

Further testing shows that Proguard is likely recognizing "@interface"
as a distinct token internally. This became apparent because
"@interface.something" at the beginning of a rule is interpreted as an
annotation type, but "@interface" is interpreted as a class type.

This change adds a specific "@interface" token to our parsing, creating
similar behavior to Proguard itself.

Fixes: 158189488
Test: Included
Change-Id: Ifb45277db15769f4ccbbd8b3f7ba74fbb0ca0b30