Use text instead of PSI to create new files
We don't need to use PSI to create new files from code blocks, we can just copy the text. Also, we were adding each fragment individually which could lead to confusing results. For example, if the snippet was a class followed by a function, the class would be added first, and then the function would be added in the "smart" place for a file containing a single class, which is as a member of that class. But this is likely not the intention when creating a whole new file for the snippet.

Change-Id: Ib28d951e656c59a050eae517c8425f31fb2ff0d6