sml: Pass errors in-channel
Currently, if there are exceptions when communicating with
the backend, they are simply thrown by the lower layers, and need
to be handled at a higher layer. We weren't doing this, and also,
it is hard to do this when spanning multiple layers of the flow.

This CL communicates the exceptions as regular messages in the
flow. As a result, errors will be automatically surfaced to the
user as just a regular chat message.

Fixes: 277347396
Test: added
Change-Id: I011ed86809a637eae0bfae44f252bf2c89e91d33