Refactor StudioProfilersWrapper
Most of the functionality in StudioProfilersWrapper will be different in
task-based UX (i.e. the window title will never change and the icon will
depend on a different aspect change), so StudioProfilersWrapper is being
hidden behind the StudioProfilersTab interface and the 2nd tab in the
task-based UX will have a different implementation for this interface.
Also, StudioProfilersWrapper is renamed to StudioProfilersSessionTab
to be more descriptive.

Bug: 270211065
Test: covered by existing tests
Change-Id: I226373d35a97b96f92e8fb61bc4f7e651b676af8