sml: hook up verified responses
This CL wires up the custom response received from the
backend to the verified responses in the UI. This is a fairly
minor change, but spread across many files:
 * AidaModel recognizes the custom response, and creates
   an SmlConversation.TextChunk with verified=true.
 * The FragmentsToBlocks transform needs to pass these
   verified chunks untouched.
 * BotModel needs to be able to persist the verified bit
   as part of its proto based database.
 * BotResponseModel also needs to model this verified bit,
   and the TextParser that creates it needs to pass this
   bit untouched.
 * The BotTimelineController emits a new type of event to
   the UI on seeing a verified message.
 * Finally, the UI layer, needs to show the verified chunks
   panel upon receiving such a message.

While creating a prompt for a subsequent conversation, verified
responses are not included as part of the prompt. This is an
arbitrary decision that probably needs further thought.

Fixes: 277100858
Test: BotModelServiceTest
Change-Id: I3f1b7e5396fb349a3dc0bf7e7931d5ff26442a83