Autoupdate of 2 submodules
*tools/adt/idea updated to 76cf8e7c5b0403bab83ed8eb8081771e1fd15605
Updated default Wear API when creating projects

The default API was set to Lollipop (21). The value was lower than the
min offline API level (23, now 25) for wear. This CL updates the default
 API to
N_MR1 (API 25) which corresponds to
the minimum version required by wear templates.

Bug: 275037433
Test: FormFactorTest
Change-Id: I1f56c7824b3609f0a852525d4b3900fcf1157e42

*tools/base updated to ce59b19525c1a6df067afc0c3148a54f52a992b4
Update Wear API levels

The minimum API level is updated to 25 which is a supported version. It
also corresponds to the minimum
version required by wear templates. Version 23 is no longer supported.
The maximum API level is updates to 30 which corresponds to the maximum
version supported by Wear 3.

Bug: 275037433
Test: SdkVersionInfoTest
Change-Id: I2f40dd3332b1e95ade100e0d5cf915ce00af4c02