Autoupdate of 2 submodules
*tools/adt/idea updated to 88558ed18dfd72ff338b6b8eee43fb96b06ca303
Adjust callers of MavenRepositories for use of Component

Bug: 242691473
Test: existing (no change in functionality)
Change-Id: I366a87038c8627d0b499e91c8b92adde12e89b4a

*tools/base updated to fd7fdd525a062b61859669d6f8a3621d3ef2fdc8
Use Component class in MavenRepositories

Various maven repository query operations are working with explicit
versions and therefore should be using Component objects, rather than
GradleCoordinate with its somewhat ambiguous handling of the revision
field (is it a user-specified Rich Version, or a single version of a
software artifact that exists).  Adjust callers appropriately.

Bug: 242691473
Test: existing (no change in functionality)
Change-Id: I3f9873c74c0b9327b1adba149454abf605c8f373