Move common parameters into FakeToolWindow
In game-tools we use FakeToolWindow as a replacement for Android
Studio's / IntelliJ's tool window. The two key responsibities are to
provide a project and to contain a JComponent.

This CL moves the common set-up values inside of it so that someone can
use it after constructing it via its constructor.

Change-Id: Idb411ff38f7589c2902a40a8d4c9c9e8f146d8f5