Look at other packages when deciding which one a reference points to
When deciding whether/how to qualify a reference in a code snippet, we
now keep a list of all the possible packages. We prefer the one, if it
exists, that is used elsewhere in the snippet.
If the snippet uses packages from compose, we also prefer the one, if it
exists, that is from compose.

Test: testAmbiguousAndroidxNames
Change-Id: I60b8dd0c2b0150a9a8e04507021f654d419a6b47