[ASfP] Destroy entire process tree after cancelling sync
Sometimes the Soong lock is not properly released by sync processes and
can potentially lead to situations where users are unable to manually
invoke Soong. This is especially prevalent when users explicitly cancel
sync, but can also occur when the IDE is closed. In addition to
destroying the entire process tree on-cancel, we also do so in a finally
block to ensure proper clean-up.

Fixes: 269771694
Test: Manual
Change-Id: I6ab155cafecfef8516db4353a6d9a705175e4324