Autoupdate of 2 submodules
*tools/adt/idea updated to 554edb0fdb5d37419ef5aef120f6ccda849613a5
Remove SwappableProcessHandler

...and replace with DeviceAwareProcessHandler

SwappableProcessHandler is not used for AC and ACC anymore, plus it had
misleading name as it used not only for swap related handlers, but for
all. We also don't need to store executor.

Check for DeviceAwareProcessHandler using user data instead of instance
of, because there are process handlers we don't create, so we can't
always extend DeviceAwareProcessHandler

Fixes: 274103920
Test: AndroidProcessHandlerTest + run tests
Change-Id: Ie868421f59910e160bf3e8c7cee90ea17b547cbf

*tools/vendor/google updated to 263129c4e8ebc3f100df44177402da3e6e8b1c35
Remove  SwappableProcessHandler

...and replace with DeviceAwareProcessHandler

Fixes: 274103920
Test: AndroidProcessHandlerTest + run tests
Change-Id: If566bbc0d5af4a43a8ef7b9af10fbca7db67877e