Add documentation code action
This CL moves the "Paste to Bot" action into a new
editor context menu, "Ask Studio Bot", which only shows
up if you've logged in and have completed the StudioBot
sign in flow.

The context menu contains two new actions:
* "Explain Code" (which is the old PasteToBot action)
* "Document Code"

The second action will ask StudioBot to document the
code. Then, when the reply appears (as a code block
with new comments inserted), there's a new Code Action,
"Document Code". When invoked, this will merge the
suggested comments back into the editor.

These code actions also ask for explicit permission to
transfer the editor code to the bot (with an ability
to ask it to stop confirming from now on.)

Test: Unit tests included
Bug: Not filed
Change-Id: I3a7a2232ee2c360063119dca538ceee068d0b4d4