Slash command for saving conversation as text, md and html
First, this CL adds a /save command which saves the conversation as
markdown, text and html (and puts it on the clipboard).

If you supply a filename argument, it will save it to that file as
markdown (or if the extension is .html, as HTML), otherwise it will pick
a new name on the desktop and save it there (just like the /screenshot
command). You can also supply an argument, --last 3 or --last=3 to for
example only include the last 3 question and answer pairs. The HTML
report will include avatar and verified icons; you can skip this using
--no-icons, as well as editor syntax highlighting colors in code
snippets, which you can skip via --no-color.

It also updates the "open" command to allow the special argument
"manifest" which will find and open the primary manifest in the source
editor, and the special argument "test" to jump between the test and its
tested class, updates the /split command to understand "down", and fixes
a bug where "debug test" would throw an exception.

Test: Manual
Bug: 278264472
Change-Id: I632c0e555bd1eb5dadb2137f7604bdcd6c2d86e1