sml: Use AtomicLong to generate message ids
We were relying on System.currentTimeInMillis(), which isn't granular
enough if messages come fast enough, e.g. at the time of loading
history. There's a suggestion to use rpc ids in the bug, but rpc ids
aren't guaranteed to be unique and we also need an id for the
question before any RPC has been sent.

Test: existing + manual test that messages loaded from history are
  cleared when the # of displayed messages exceeds 8 (max limit).
Bug: 274089245
Change-Id: I72c3c156717f637aaf0d0caa60205c0743130daa