Autoupdate of 2 submodules
*tools/adt/idea updated to 4bf1aa60d2eeb26a2a315878d854f5e71edf4cca
Poll for connections in configuration manager

With this change the UI's connection dropdown is able to show available
apps users can connect to. In addition, the proper AppInsightsModel is
emitted based on the login status of the user.

Note: the name of the app in the dropdown is not correctly formatted yet

Test: added new test.
Bug: 271919516
Change-Id: I969f0bd6f0e9e4a1f63a45b9f83ab361e980b592

*tools/vendor/google updated to 0f815ac500905422a466f0de811775c4a0eccf3a
Rename AppInsights classes to Crashlytics

This cleans up the firebase module and renames any classes with the
prefix AppInsights to a Crashlytics equivalent.

Test: n/a
Bug: n/a
Change-Id: I3a146a4118fb2c6e4cc21efb70ce5fa288fce859