Clean up ComposableItemPresentationProvider
ComposableItemPresentationProvider needed to modify function
presentation in some cases, and defer to the "standard" presentation
otherwise. To do so, it had to mimic the standard presentation since the
Kotlin plugin didn't allow overriding some (but not all) presentations
for a given KtElement type.

Now that https://github.com/JetBrains/intellij-community/pull/2209 is
available in the studio prebuilts, that workaround can be removed and
the provider can just return `null` as expected. The plugin now will
keep checking for more providers in that case and will fall back on the
default itself.

Fixes: 255608890
Test: ComposableItemPresentationProviderTest
Change-Id: I393335a9b60b6a8dd33cc3f55e5ba8fe5799279f