Autoupdate of 2 submodules
*tools/adt/idea updated to 3a2d8187851584bbe455b20b8f4a00dced9b835f
Remove display mode param from power request

Before, the way in which power data was processed
was determined by different SQL queries in the trace
processor daemon. The query itself was determined
by the display mode feature flag for the power profiler.
However, because data processing is now going to be done
on studio-side, we can use the same SQL query for all
display modes. This shared query being essentially the
raw data.

Thus, this change removes the setting of the display
mode parameter when creating the power data request.
A corresponding change will set a default SQL query for
power data in the trace process daemon.

Bug: 254116554
Test: updated
Change-Id: I6ee5f98dd81851127382b6395ba33a13ae1d93e8

*tools/base updated to ea18e46217c366894053c22db62344945f66ba9c
Set default power data processing query

Before, the way in which power data was processed
was determined by different SQL queries in the trace
processor daemon. The query itself was determined
by the display mode feature flag for the power profiler.
However, because data processing is now going to be done
on studio-side, we can use the same SQL query for all
display modes. This shared query being essentially the
raw data.

Thus, this change removes the display mode field
from the PowerCounterTracksParameters proto. It also
sets a default SQL query for power data in the trace
process daemon.

Bug: 254116554
Test: updated
Change-Id: I014dca2ac5238fcea466948a47a70bce24aa2d34