Add missing write actions
ag/22562286 removed a write action. Usually the contents of this action
are invoked from the EDT. However, when they are called back by a dialog
a write permission error is thrown. This CL fixes this, but it also
means that if part of an action is done by a dialog callback,
the result of the action cannot be undone in a single undo action.

b/276644134 has been re-opened

Bug: 276644134
Change-Id: Icfe2b588dab314cc954e74806d19c3874c1b828e