Don't qualify functions
It's very hard to tell if a function call is an extension function or
not. Some function names correspond to both top-level and extension
functions. Extension functions can be invoked without a dot preceding
them, e.g.
in the test code snippet.

As a workaround, this CL turns off qualification for all lower case
names, and instead opts to add/keep their imports.

In the future, if we can associate every name in our dataset with some
metadata about what it is, e.g. a class, top-level function, extension
function, etc. we might be able to do better than this. But ambiguity,
especially in the context of code snippets where many names don't
resolve, will still pose a problem.

This CL also fixes some small other bugs that were found when modifying
the test cases. The test cases cumulatively cover these bugs. For
example, there was a missing . in some of the added imports, and the
same import could be added to the same snippet multiple times.

Fixes: 278602799
Test: tests in KotlinParserTest, especially testExtensionFunctionsAreNotQualified, which demonstrates the original issue
Change-Id: I699b4fafa07852bfda5c1710ba20b38c7922bcdc