Autoupdate of 2 submodules
*tools/adt/idea updated to e2ba1df211c992d000d1e971454dc01cf261e05a
[Vitals] Parse report text to structured stack trace groups

We are not giving perfect solutions here but just
best efforts. If a trace is parsable, we will be
able to collect file name, line number or so for
annotations as part of the App Insights (Vitals) support.

For printing traces in our console, concatenating
everything like we do for Crashlytics will still work with
this change, even we might have imperfect trace groups
extracted.

Bug: 265153845
Test: added
Change-Id: Ifbfc3f9b1947a4d6500e2bc6d25e5462fef5f0bb

*tools/vendor/google updated to f7cb89bd967d138b5a5480c165e9f77d2f1413d5
Set raw strings to accommodate changes for Vitals

Basically, we concatenate strings at an earlier stage,
with this change, we do it when building up the data model,
instead of when printing out text (traces) in the console.

Bug: n/a
Test: n/a
Change-Id: I10c985ff55ab14bfe073da30714fa4120721dd88