Fix complication configuration not showing slot
... options when run from the gutter.

This change restores the change from ag/22650181 that was reverted
due to a flaky test, without the flakiness. The flakiness was
caused by some tests running assertions before the model was updated
from the EDT.

The slots not showing was due to the EDT thread being blocked for
non-modal events preventing the "onFinish" method from being called
until the dialog was
closed. This is because the dialog does not yet exist when we register
the background task. The background task gets registered with a
ModalityState.NON_MODAL meaning the "onFinish" method gets queued on the
non-modal queue which gets blocked when a dialog is shown. This is
because we enter a secondary loop when showing the dialog[1]. The
"onFinish" method was called when closing the dialog as that's when the
secondary loop is exited and the blocked events get processed.

This issue was not occurring when editing the config from the toolbar
because the
editor content was wrapped in an existing dialog meaning the background
task was registered with an ModalityState that was not NON_MODAL.

This CL also fixes an issue relating to the complication slots not being
 repainted properly when a second slot is selected or deselected. I'm
 not sure why that was occurring.

[1] docs.oracle.com/javase/7/docs/api/java/awt/SecondaryLoop.html

Bug: 214240163
Test: AndroidComplicationConfigurationEditorTest*100 + manual
Change-Id: I6025644ec95b21035f596c7b4fd0668933801e08