Autoupdate of 2 submodules
*tools/adt/idea updated to 394b68a67b1e49743d493a8d51b09af59e392dbe
Add an "Explain This" extension and hook it up in Studio

First, this topic adds a new extension point, IssueExplainer, which
all the various Android plugins can depend on. The service allows
clients to request details on a specific error.

In the LogCat window, you can right click on log entries and
crashes and ask for an explanation. (For crashes, it will
identify the root cause and only ask about the 5 deepest
stackframes along with the logging tag.)

Bug: 274933635
Test: ExplainLogcatCrashActionTest.*
Change-Id: Id2d2693d33a42f156e6502f9c838d968b498a243

*tools/vendor/google updated to 3b3e88d793f5926bbbfba9b88bd37f277a07a0a0
Add "Ask StudioBot" extension & hook up to a few places in Studio

This topic adds an issue explanation service to the Android plugin,
which is generically only referring to "explaining" issues since
that branches is visible externally.

Here in StudioBot, we map the explainer functionality to Ask StudioBot.

This CL also has some hacky workarounds to handle a couple of scenarios
I ran into:

(1) The action to open StudioBot on a query would often result in
    the query text missing if the StudioBot had never been opened
    in the current IDE session; it looks like the tool window
    construction isn't quite done by the time the flow was sending
    the text

(2) In some cases keystrokes from other parts of Studio would
    be processed by StudioBot right after the focus transfer (for
    example, if you use Enter to select a quickfix to ask StudioBot,
    that gets executed on key down; which would transfer focus, and
    StudioBot's text field would see the enter keyReleased event,
    which it then took as a sign to submit the query).

Bug: 274933635
Test: manual
Change-Id: Id5090c1060dd88d0c821312541269635d501fafd