Autoupdate of 2 submodules
*tools/adt/idea updated to 80557b3f0e9c795023c3a2f8905ee03b4397d24e
Simplify tool window wrapper

When game-tools tried to implement the tool window wrapper, it
conflicted with JFrame. Upon looking closer, there were a number of
methods defined on the interface that were not used, allowing the
interface to be simplified.

Change-Id: I8d234acb5c16c75795d80dca37bbf4f26e3de7d5

*tools/vendor/google updated to c0bfb08f803eb8593bd947d2c7cf07a365855fcc
Simplify tool window wrapper

When game-tools tried to implement the tool window wrapper, it
conflicted with JFrame. Upon looking closer, there were a number of
methods defined on the interface that were not used, allowing the
interface to be simplified.

Change-Id: I8d234acb5c16c75795d80dca37bbf4f26e3de7d5