Improved XML manifest error recovery
The LLM can often produce XML fragments that are not valid XML.

This CL attempts to fix a number of these scenarios:

* Missing closing tags, or worse, the closing tag is listed
  as a second open tag (this is the scenario in issue 278935773)
* Unexpected tags
* Omissions/ellipses (like "...") in the middle of an attribute
  list where no text is allowed (and it's probably not to be
  included anyway)
* Missing the correct parent nodes (such <style> elements, not
  including the <resources> parent)
* Missing namespace declarations (e.g. an XML snippet which
  references android: but doesn't bind the namespace anywhere)

It now has a simple statemachine which looks through the XML
fragment and keeps track of the current tag stack, and as it visits the
XML inserts missing closing tags as necessary (or drops unexpected
closing tags), as well as removes ...  where that makes sense (e.g. from
text, if only surrounded by whitespace, or from attributes (not
attribute values)).

Fixes: 278935773
Test: Unit tests included
Change-Id: I5ab93d15f1fe3622c85f1b5992678c8d786e42f4