Don't add merge manifest actions for trivial snippets
For a question about the internet permission, I got a response where the
model had created separate code block responses for the reference to the
filename (AndroidManifest.xml) and the <manifest> tag. These ended up on
separate lines with editor syntax highlighting, which I guess is working
as intended, but we also added a Merge into Manifest action for
"<manifest>" which doesn't make sense. This CL looks for trivial tag
references and skips adding the action for these.

Test: Unit test included
Bug: None
Change-Id: I89a347cf6efe62d4c1f108335d368c3a1a89bb21