Move element finding out to a shared location
The slash-command support had a number of utility functions to find the
most relevant elements:

* What's the current file?
* What's the current module?
* What's the best PSI element match for a given class name?
  Method name? Fully qualified name? File name?

This was based on a number of heuristics. Typically, it starts looking
at the current editor, and if not found there, looks at the current
module, then nearby modules (e.g. source set modules in Android), then
module dependencies, and finally, the project.

Finding the current module is itself heuristics based; for example, if
you're looking at a scratch file, there's no current module, but we can
typically guess from the recent edit history.

There are also considerations like preferring source files over binary
files.

This code has now been moved into the existing parsing utils functions.

Second, this CL also update the "Insert into New File" code such that
when it picks a location to write a file, it doesn't always place it in
the same folder as the current file.  If the new file is a test, it will
go to a test source folder, and similarly, if the current file is a test
and the inserted block is not, it will go to a production source folder.
It also tries to match the package name for the current file, if any (or
if not found, for Android, picks the default package name).

Test: Manual
Bug: None
Change-Id: Ib0c3b9df971fb27eca1e203ec33121731c64e167