Add support for Android resource code actions
This CL adds code actions for Android resources:
- Insert into new file
- Merge resources (for value-based resources, like <string/>)
- Insert at cursor

StudioBot looks at resource fragments and tries to guess what type of
resource it is. For value based resources, such as strings, dimensions
and so on, it will add a Merge Resources code action. When invoked, this
will merge the listed new resources into existing resource files (if
found; if not a new resource file is created). If there are name
conflicts, it will insert the new resources commented out.

For non-value based resources, such as layouts, menus, drawables etc, it
will create a new file in the right folder.

Also fixes a few things around language recognition:
- Improve recognition of manifest tags
- Add recognition of TOML/version catalog files
- Improve C/C++ recognition (look for more preprocessor directives,
  and when found, pick between C/C++ based on the presence of
  some telltale signs of C++)

Fixes: 279071018
Test: Unit tests included
Change-Id: I3baa40be73de82a51f04af9dd3a7f39b8aad7841