Implement Wear pairing actions and status icon
When the wear pairing status changes, we need to update the table to
reflect it. Thus, we listen to status changes on the WearPairingManager,
transform it into a Flow of immutable data, and incorporate it into the
DeviceRowData.

Add a link icon to the DeviceNamePanel when a device is paired.

Add actions for pairing and unpairing devices to the overflow menu.

Bug: 276454402
Test: Added
Change-Id: If0d33b70a4477481dab5e4998528f89618896e56