Re-enable TestRClassesTest code after merge
A new assertion added with
https://github.com/JetBrains/intellij-community/commit/c052c934734d12133bddac322c39f4ea6e899c79

The assertion requires that when checking highlighting on a file in
tests, the file must be in a project source directory.

These tests were failing because the project source directories are
determined during gradle sync, and at that time there were no files in
the */androidTest/java directories. By adding test data to the project
structure in those directories, there now will be a file in each and
they will be considered part of the project source.

Fixes: 278944241
Test: TestRClassesTest
Change-Id: I1081477a009a21a10f0614213d545bada1f53fd5