Make Dagger gutter icon providers single extension
The user doesn't know (or care) that there's a new implementation of
Dagger gutter icons for the most part. Combining the two implementations
into one extension that wraps both allows there to be a single entry in
gutter icon settings that controls turning both on and off as a single
unit.

Fixes: 278896170
Test: DaggerRelatedItemLineMarkerProviderTest
Change-Id: Ib19e23f6d363f77a1961fa729be361d6e548a8c7