Decide running Visual Lint from LayoutlibSceneManager
At the moment, whether to run Visual Lint is decided at the surface
level. For Compose, there can be multiple previews on a given surface,
so we want to be able to have more granularity in deciding exactly which
previews should be analyzed.
In addition, this introduces two different modes of running Visual Lint.
One which creates additional models in the background to analyze, which
corresponds to what is happening in the main layout editor surface. And
one which simply analyzes the current preview, which is what Layout
Validation does (and what we are planning to do for Compose).

Bug: 274150043
Test: Test added
Change-Id: I6fa74101930d6ebcf492cd5641a1e8d759f6f869