Autoupdate of 2 submodules
*tools/adt/idea updated to b0a53ed55ce870caeca9d9aedb94355fb9d3d041
Display Kotlin sources under Kotlin node in the Android view

Make kotlin sourceType be mapped to both java/ and kotlin/ folders,
do this by allowing source roots intersection for java and kotlin
sourcetypes.

Add filter to java and kotlin nodes with the following logic.
For Java node we want to remove kotlin files but only if kotlin node
would also contain the source this file belongs. Otherwise the file is not shown anywhere and we want to avoid that.
For Kotlin node we want to show all the files under purely kotlin
sources (keep as it was before) but additionally show kotlin only files from java source folders.

Bug: 273905575
Test: manually verified, updated tests snapshots
Change-Id: I2b26f7c2878853aa77ffb7f31ab44f02392f394f

*tools/vendor/google updated to 2b908da7e303b19d1d266fdee8e8ca374309d81b
Display Kotlin sources under Kotlin node in the Android view

Bug: 273905575
Test: manually verified, updated snapshots
Change-Id: Ic97237b8392e4ac6f3358d77d3ff1dc738e8477e