Autoupdate of 2 submodules
*tools/adt/idea updated to 33212c7193e49988aa4e6354837e063d9103db33
Remove remaining uses of GradleVersion in android core code

This isn't actually such a radical change as it sounds, given that the
replacement is also explicitly gradle-specific, albeit in its package
`com.android.ide.common.gradle`, so there is still work to do to
abstract away the build-system-specific nature of some of the
operations.  But this at least gives some reassurance over
implementing correct semantics for at least *one* build system.

Bug: 242691473
Bug: 272742663
Test: existing (no change in functionality)
Change-Id: Ife6593f05735ef8e6bbe1a76a42330702d4e01bb

*tools/base updated to 15535d1857eac272bef3f76abd4a70c7986cdad0
Add a way to find the kind of preview for a Version

For many (but not all) version conventions, the non-numeric components
in a version indicates in some way the kind of pre-release version it
is: not just the treated-specially-by-Gradle versions such as "dev",
"rc" or "final", but also user-specified strings such as "alpha" or
"beta".  (Some other non-numeric components might have little or
nothing to do with anything related to this meaning, unfortunately).

Some areas of Studio might wish to display this for the user.  We
cannot do better than an approximation to this (in principle users can
use arbitrary non-numeric strings for any purpose) but we can do
something that is reasonable for the common case.  For cases with more
than one non-numeric part, we return the string representation of the
logically-least part, reasoning that the user will usually want to
know the maximum degree of previewness encoded in the version.

Bug: 242691473
Test: added
Change-Id: I37b605c78932f7e1d239161e2d6b514b3702af52