Autoupdate of 2 submodules
*tools/adt/idea updated to 46065287d24656b12566dfc09cd9fc7ae78d0fbc
[Vitals] Refactor code around the insights annotator

Basically, an "InsightsProviderKey" field is introduced in
"AppInsight" to denote the source of insights, e.g. Crashlytics,
Play Vitals, etc. Then we can simplify the structure of the
insights we pass around. And we won't group insights by its
provider until the we render items for the popup of the gutter
icon action.

Bug: n/a
Test: existing
Change-Id: Ibb95abf344ef4fc103f83c077fa7ec84ab4fde10

*tools/vendor/google updated to 0d96ace5c38f8b7da5a31bf1e7216482d8bcce9c
Minor update in Crashlytics code

Bug: n/a
Test: n/a
Change-Id: I57869b5c4c01d3ad3b80435ba35b92d91468ca30