Autoupdate of 2 submodules
*tools/adt/idea updated to 1f1d2fa80c5d248621eb8c35b7cffa00513f6dd2
Cleanup ResizableBuffer#afterChannelRead

ResizableBuffer#afterChannelRead is used to return the ByteBuffer with
the position that can be set to zero or to the position marked
by a call to [forChannelRead]. Most callers should just use a marked
position (which is the default).

Bug: n/a
Test: n/a
Change-Id: Ie49d4f7c986859640475d602c760c8fe8b7d844f

*tools/base updated to 8a23d7bb43feff7be2d507a2d89ba92297a1f8bc
Cleanup ResizableBuffer#afterChannelRead

ResizableBuffer#afterChannelRead is used to return the ByteBuffer with
the position that can be set to zero or to the position marked
by a call to [forChannelRead]. Most callers should just use a marked
position (which is the default).

Bug: n/a
Test: n/a
Change-Id: Ie08f49f5e99b86b9dc7ca298b0f1c48c040367d0