Autoupdate of 3 submodules
*tools/adt/idea updated to c17a0221eda410be0bca0edc30e4ff42de9f2209
Adjust callers of SdkMavenRepository

Bug: 242691473
Bug: 279886738
Test: existing tests adjusted
Change-Id: I512f1a5ef959afbef622c9a58ab8b90b2a0d5c35

*tools/base updated to 9cb3ff411c07ff32f47cf2f3e1637dbef37ccfb8
Use new Gradle representation classes in SdkMavenRepository

Replace uses of GradleCoordinate with more specific types.  The change
in GradleDetector is to support the propagation of modifying callers
in the IDE code.

Bug: 242691473
Bug: 279886738
Test: existing tests adjusted
Change-Id: Id1e858fa3bc36a0c7c16fea1bfda72a67e8d9502

*tools/vendor/google updated to d5389971b29b7481e297a25305fb2c80b8ff9560
Adjust caller of RepositoryUrlManager

(From the update to use Dependency at the interface rather than
GradleCoordinate.

Bug: 242691473
Test: existing
Change-Id: I3bd710f6c2b8b2dce18a4980835d0fabe59d59d9