279973496: Fix module selection bug
The "select module" dialog used by the add dependency action limits its
options to "real" modules; ones with content.

However, if you don't have any open files, or if the open files aren't
in a module (and in the test scenario for this bug, only a scratch file
was open), then the current module returned by the current module lookup
function can be something else, such as the root module.

The selection dialog initialized the selection to the current module,
but if that's not one of the options, this threw an exception.

Test: Manual
Fixes: 279973496
Change-Id: I6d0a4eb9ea66f92ab380452dd85d736d936fb9f9