Make Compose-related tests non-parameterized
We made a lot of tests parameterized when migrating the Preview and
Composable annotations to different packages, to make sure we verified
the behavior when both new and old packages were used. Now that the
migration happened a while ago, we don't need to keep running the tests
using different package combinations.

Bug: N/A
Test: compose-designer and compose-ide-plugin tests
Change-Id: Ic1e65a79fc0fe9d58901ca15338702ee4c2d163a