280076019: Java or Kotlin Library is listing in select module dialogue
The bug is that we shouldn't offer any modules in the module chooser
that don't have a manifest file for the code snippet to be merged into.

This CL updates the module selection utility to take a module filter,
and we use that one to filter the list down to modules that have
manifests only (and if there are no module matches, it will show a
warning dialog stating that no modules contain a manifest to merge
into).

Test: Manual
Fixes: 280076019
Change-Id: I29669b5c2c12756931524ab620b7d284904f4c62