Allow verified messages to replace generated responses
This CL adds support for verified messages coming back from
the server to carry extra metadata, such as whether the
message should be shown in addition to (the default) or
instead of the generated response. It also adds some basic
support for gracefully handling future metadata.

This makes the assumption that verified responses will
arrive before generated responses that they're supposed
to be replace. This should be the case "always"; if we
want to worry about delayed verified responses we'll need
to work a bit harder.

Test: Unit tests included
Bug: N/A
Change-Id: I4accf06bb0d0da3e8aeb6312636de1785d193a4b