Update plugin metadata
This CL updates the plugin metadata, adding a plugin icon and a new
description that better represents the plugin capabilities.

The changes are also synced to the IntelliJ plugin, along with other
tweaks that were not brought over (e.g., the plugin ID change).

Lastly, the commit message settings in the IDE settings have been
aligned to the main Studio project, for improved consistency between the
two projects.

Bug: none
Test: manual
Change-Id: I12eadf596866d91ef05aab42842bf837f4696e11