Fix rendering of fallback user profile picture
I658ec090b51f8 deleted the file loginAvatar.png. This file was
used as a fallback profile picture if one is not available. As
a result, this flow is currently broken.

This CL fixes it by redirecting the code to use
GoogleLoginIcons.LOGGED_IN_FALLBACK_USER_AVATAR in this scenario.

Fixes: 280213828
Test: manual, by forcing the profile pic to be null.
Change-Id: I87f394e3029bae286a05f31b994c7628feb3b7b5