Prepare infra for geolock checking and wire it up to the UI
As far as I know, there is no direct way to tell if we're being
geo-locked. The only thing we can do is to try and call some API and
see if it blows up, than hope that it's not blowing up for whatever
other reason (e.g., network error).

Given that limitation, this is a minimal implementation that does what
it can to try and obtain and display this information anyway.

We have two GeolockChecker implementations:
 * ModelBasedGeolockChecker, that checks by trying to call a cheap API
   on the active model backend, and seeing if it fails
 * TestGeolockChecker, that is only activated (in internal mode)
   manually, and allows for testing the behaviour of the plugin in all
   the different scenarios

If we detect a geolock during the onboarding first step, we prevent the
user from proceeding further. After they clear the onboarding, for now
we don't perform any further checks.

Bug: 280060093
Test: manual
Change-Id: I7dbac884939229f63bc79e4f30164d77112118a2