Autoupdate of 2 submodules
*tools/adt/idea updated to ba68d009d8892729a2e8a26771b948a23e82786e
Profiler window supports 2 tabs

When the task-based UX flag is on:
- the Profiler tool window initializes with an (almost) empty home tab
- clicking the button on the home tab opens a second task tab which
  displays the NullMonitorStage
- in the future, starting a new Profiler task will populate the task tab
  with actual profiling data

When the task-based UX flag is off, the Profiler tool window should
behave exactly as it does today.

Bug: 277797525
Test: added TaskProfilersViewTest
Change-Id: I39e5ba6c1f6cd116d4c9bc912a0480d0069904cd

*tools/vendor/google updated to 61a13ffe955be145dcae88e24a0e73bd47228130
ToolWindowWrapper#getContentManager placeholder

In the new task-based UX, the Profiler tool window will have at least
two tabs, so the ToolWindowWrapper needs to provide a ContentManager. We
haven't finalized the CUJ yet, so this is just a placeholder for now.
The getContentManager should never be called when the task-based UX flag
is set to false.

Bug: 277797525
Test: covered by existing tests
Change-Id: I64f955ae76c2706884e5c7faa84dd2ee8a7342c7