Show name choosing dialog before scope choosing dialog
When creating new functions from Kotlin snippets,
we now first show the dialog to choose the name of the
function and which informs the user that the snippet
couldn't be added at the target location, instead
of first showing the scope-choosing dialog.

Showing the scope-choosing dialog first is confusing
if you don't know why it is showing up or why a function
is being extracted. Now the purpose is clear.

Bug: none
Test: manual
Change-Id: Ie705dbd6c6cda7c0def7eac64a812d2ff05096d7