Implement Java code actions
This CL enables the "Insert at Cursor" and "Insert in New Java
File" actions for Java code snippets.

The Java actions and parsing are more limited than their
Kotlin equivalents. The only snippets we consider for actions
are those which contain either

 - Import statements and a class
 - Exactly one method
 - A list of statements that can go inside a method body

For all other snippets, parsing will be difficult to get right,
and so we do not make any actions available.

Also, the actions themselves are simpler. There are no scope-choice
dialogs, and methods always get placed in the innermost class
containing the cursor, in the single class in the file, or in a
new class if the file doesn't have one. The same goes for classes.
We don't add any extra keywords.

Qualification and smart-imports will be added in a follow-up CL.

Bug: 274686803
Fixes: 280312754
Test: JavaActionProcessorTest
Change-Id: I3c7a4413881f8832bb7dd70371b4b628dbd3bf0b