BotToolWindowController: clear timer listener on disposal
Currently, we create a timer that sends an update to a merging
update queue. But since there is only 1 update that is sent (from
the timer event), there is no need for a merging update queue.

In addition, the timer wasn't getting disposed. The listener
was holding on to a reference to the panels and indirectly to
the project. We now register it to be disposed along with
the tool window.

Leaving the bug open as I still see some objects being
retained.

Bug: 280301673
Test: manual, close & re-open the IDE thrice
Change-Id: Id08276c9696903189b4e4edec6f9d35353e4eb48