Autoupdate of 2 submodules
*tools/analytics-library updated to c4ade797f2e2f473fdcdd9ab6e7519094d38c641
Add flag for api modeling and global synthetics

Bug: 254510065
Test: existing
Change-Id: I7e57326bcf95522635fbc6945cd9a8aac2d3c265

*tools/base updated to 40de9fef1be5d9ce1f129a76c2d0757b4118e310
Mitigate perf regression caused by globals for debug build

This CL introduces an optinal boolean flag to control both api modeling
and global synthetics. When flag is set, we fully respect the flag. When
the flag is not set, which is the default mode, api modeling and global
synthetics are enabled in release build. For debug build, api modeling
is disabled and global synthetics is only enabled when java language
level is 14+.

In this way, release build can benefits from api modeling and the debug
build can avoid unnecessary performance penalty. To address the
performance issue completely, b/279930074 is filed.

Bug: 254510065
Test: GlobalSyntheticsFeatureTest, etc
Change-Id: Ie813e4d16a340f28ed3fb3fd2f1dc666738ebc89