Disable code actions for languages other than Java,Kt,XML,Gradle
There is currently a default implementation for many other
languages by way of DefaultParser and DefaultActionProcessor,
but these implementations haven't been refined for each
language and so are prone to bugs.

We should return to these languages when we are able to
dedicate time to making sure each one has reasonable
behavior.

Change-Id: Idd6e2053b3d048ba463d62f111c75313ff3daefa