Improve chunks layout logic
The chunks' layout is now specialised, simplified.
This CL prepares the road for making messages take up the full width.
To enable that behaviour you just need to flip the flag in MessageLayout
but in the meantime we're already forcing all messages with any code
chunks in them to take up the full width.

Message chunks size is also capped to a maximum value to keep them
readable, as very long text lines are difficult to read.

Bug: 280548773
Test: manual
Change-Id: I7fa946a4d059a0df0fa528d8d049944c729facbf