278743800: Getting IDE error after clicking Merge into manifest button
Optimistic fix for 278743800. The document support is throwing
exceptions about \r\n appearing in documents (these are normally
removed on file read and added back on file save). I can't reproduce
or verify this outside of Windows, but from the stackframes it
looks like this should fix it (passing in null here causes the
printer to pick up the platform line separators; instead, we just
want the unix one). I'm reimaging my Windows cloudtop instance but
that will probably take too long so putting in a fix to to see if
we can get it verified in the meantime.

Test: Existing tests verify that things are still working. Don't
  have a way to force the Windows behavior.
Bug: 278743800
Change-Id: Ib1f7a9fc6df74897c5ec25b7f3aa78dcaea8b136