Temporary fix for companion object snippets
When inserting, we now check to see if a snippet is a companion
object declaration. If it is, we only add it if the cursor is inside
a class. If not, we show a message.

The real fix should involve refactoring CodeFragmentType to carry
more information about what snippets contain. For example, code snippets
containing both a companion object and a function will still have
broken actions. Hopefully that type of snippet is rare and will only be
generated by people trying to break the actions.

Companion objects are problematic because they can't be parsed the
same way as other declarations, because they don't belong in the top-level of a file.

Bug: 280684131
Test: testCompanionObject
Change-Id: I653098a3c06e7cfeefdd5c1e16abdf6a402d6b4e