Fix minor issue with timeline empty state
Without a minimum height, the timeline empty state would get clipped in
odd ways when the toolwindow wasn't tall enough.

Now we use the preferred height as min height, because since the
contents are dynamic, we can't really set a hardcoded min height.

This seems to work well, especially in conjunction with a smaller
vertical padding around the empty state, which saves us some more height

Bug: none
Test: manual
Change-Id: I7ee79007f9425053fba162ba1812f273317df33c