Autoupdate of 2 submodules
*tools/adt/idea updated to 73c515c0c520aaa2da5e2d474c0e9c7bff2ff1b5
Query the cacheable graph representation

This falls back to the old VariantDependencies model if the new one is
not available.

See the topic change in tools/base for more context.

Test: Covered by model integration/snapshot tests
Bug: 279570192
Change-Id: I9744ce5d5494d9c3c0b57dcdf287db335cfe6330

*tools/base updated to a1fb0c155fa520ace8fc23cf715d79d346ef1aa0
Switch to a cacheable graph representation

Adds a new version of VariantDependencies with an adjacency list
representation instead, which makes the individual edges of the graph
globally cacheable, avoiding having to recreate them with each gradle
project instance.

This, along with the library cache
(change Ia2c774194783d41a0b93bbffeba175cc3c968785),
brings down the measured heap usage as follows:
1000 benchmark: 6842 -> 6049 MBs ->= 11% decrease

Test: Covered by model integration/snapshot tests
Bug: 279570192
Change-Id: I99a198c27945d32040eee8ac009000ee7f22d557