Correctly target the start of lambda expressions
The "Insert at Cursor" action wasn't working right when
the target was the left brace (or whitespace following it) at the
start of a Kotlin lambda.

This is because of a PSI inconsistency. For functions, loops, and other
elements containing a block bound by braces, the braces are children
of the KtBlockExpression. For lambdas, however, the braces are siblings
of the block, and children of a KtFunctionLiteral element. We use block
expressions to determine where code should be placed - we search up in
the PSI tree until we hit one (or another container type element like a
KtClass) and then place the snippet as a child of it. And so in this
case we miss the function's block, and continue searching upwards,
meaning the code ends up outside of the function.

Fixes: 281711081
Test: testMoveBlockToStartOfLambda
Change-Id: I08969249bd1f389a451209b584635c441e51c628