Capture histograms with soft references minimized
This makes use of the flag SoftRefLRUPolicyMSPerMB, setting it to 0
to try and reduce the impact on soft references on the histograms
captured via jvm tooling. This is an experimental mode which might end
up helping us get much faster measurements but the value stability is yet to
be seen.

Bug: N/A
Test: Test change
Change-Id: Ia91850cbb4d790b01c37b89c617295dc1eae7135