Shows a render error if PreviewParameter has no values
In case providerCount had a zero value because of an empty sequence it
showed no Parametrized Preview producing no Previews nor error messages.

This change fixes such behaviour as we still want to communicate a
warning in the Preview, in case the providerCount is zero we still show
a ParametrizedComposePreviewElementInstance with 0 index showing a
render error that an empty sequence has been used.

Test: ParametrizedPreviewTest
Fixes: 266920586
Change-Id: I68aa964ca9c0d1f4fa031b0850d7d88f2f1c6dbb