Autoupdate of 3 submodules
*tools/adt/idea updated to 8e9975b6d6dcad534539558eb406c5dbc7463a2a
Rename a RepositoryUrlManager method

Make it clear that it returns something which can be used as an
Identifier for a Component.  Rename some variables at use sites too,
and remove one or two needless uses of GradleCoordinate.

Bug: 279886738
Test: existing
Change-Id: I915bd17b7d463ab1843c4016b7eca1424ed5bc61

*tools/base updated to 3279faaa83bfaf8289044b713d495f2ae76f82b4
Include plus-suffixed Gradle version string in the lint fix

This prevents the implementation of the fix having to reparse it from
the coordinate string.

Bug: 279886738
Test: existing
Change-Id: I4a8d52ebbccd61be976e4459d14f8146fc3c5245

*tools/studio/google/cloud/testing updated to e26eef45eed422b5ea16ac747500582375600523
Update for RepositoryUrlManager method name change

Also remove a GradleCoordinate use.

Bug: 279886738
Test: existing
Change-Id: Ieb82118f75e0b987b480676ee13f0be0e3a6648f