Autoupdate of 2 submodules
*tools/adt/idea updated to 4716c3af7fab885eb57ad763bb77e9793d7b5d03
Make the version catalogs as an only option for new projects.

Second attempt of the same change as ag/22632478 because that
was reverted because of the breaking build.

Fixes: 278084160
Test: TemplateTest, UI tests, CreateProjectTest
Change-Id: Iae78411c6aa4748716f175cc1f1de75e7dcea02f

*tools/base updated to f1212478b50c2fa8f23d48ba4c4083d6f97c535c
Make the version catalogs as an only option for new projects.

Second attempt of the same change as ag/22632478 because that
was reverted because of the breaking build.

Fixes: 278084160
Test: TemplateTest, UI tests, CreateProjectTest
Change-Id: I28260e21bd5f5dcd5865ee581b4641659aef9f56