Autoupdate of 3 submodules
*tools/adt/idea updated to 3a55d964567bfaf8679b0620ceeedf2c6962e237
Add Cold Boot action

Bug: 276974127
Test: n/a, glue code
Change-Id: I183d8a2f9f080b7d243275f0ee4462eb7de9bcba

*tools/base updated to 3e5c258f628ec54698c90836678f2ca5958f4a56
Add ColdBootAction and remove ActivationParams

For uniformity, define cold boot as its own action rather than having a
parameter to ActivationAction. The existing ActivationAction is
incomplete, as it is lacking a way to indicate whether a particular
activation mode is supported. Cold Boot would also have a different
presentation, since its name is different. Defining it as a separate
(optional) action accomplishes this, and is consistent with the way we
generally indicate availability of actions.

Bug: 276974127
Test: added
Change-Id: I89afe511efe237329b7cfde5840f2a10870b9040

*tools/studio/google/cloud/testing updated to a0730a6803b59d082bed85094f842b58c0f8a1b4
Update signature of activate()

Bug: 276974127
Test: n/a
Change-Id: I5cb660cad29acc1603da19db515fc68949ccc930