Autoupdate of 2 submodules
*tools/studio/google/cloud/testing updated to 29c9cac361fc2b6a8956fcc3d189a8ddc86751c5
No notification when grace period is not added

Grace period is not added when the user cancels the reservation while
waiting for it to activate. Don't show the notification when grace
period is not added.

Bug: N/A
Test: Added
Change-Id: Iaa8a610956bff1dfd6fcafe42623e4ad70a233dc

*tools/vendor/google updated to 47082635adefbe86af37ec4cb96fd975eb9d40b4
Skip grace period if reservation is not active

Add grace period only if reservation is active and grace period is
requested. Cancel reservation immediately otherwise.

Bug: N/A
Test: Added
Change-Id: Ib9e4a149e199ff250dc6d1eaf0b6fdeabcd73b14