Add symbol dirs for androidTestImplementation deps
...in addition to implementation dependencies.

This change enables the discovery of symbol files for a 'lib'
module when running the instrumented test of an 'app' module
that has an androidTestImplementation dependency on that 'lib'
module.

Note that the native debugger cannot exactly know if it's
debugging an app or an instrumented test. That's why, even
if the user is debugging an app, the symbol directories for
its androidTestImplementation dependencies must be explored
and included (in addition to its implementation dependencies).

Bug: 228846664
Test: TODO, add three tests to lldb-integration-tests
Change-Id: I5f9ec9c008fabc6966b1855fd416e745c96a21ba