Allow multiple consecutive queries via code actions
State flows do internal deduplication of messages, so if you invoked a
code action on a selection, then cleared the query bar, then invoked the
action again on the same selection, it'd ignore it as the state flow
would dedupe it.

This small CL introduces an internal query representation that combines
a timestamp with the query, to prevent deduplication.

Bug: 278945924
Test: manual
Change-Id: I35deda051e98287651fa1c4bfa6c729bafc248f3