MediaControl: Extract CCCD storage out of MediaControlGattService
It is required for multiple MCS GATT service instances to have their own
unique set of CCC values stored. CCC descriptor values cannot be shared between
the instances. Since the service-owning profile knows the context in which
the service is intantiated, it is able to store the CCC descriptor values with
the proper context.

This patch provides CCC descriptor value storage for the single Generic
Media Control Service instance, stored under the BluetoothDevice.METADATA_GMCS_CCCD
metadata. Other instances will not get their values stored, and empty list
will be provided if they try to restore them. This is fine for now as
there is no support for multiple MCS instances enabled yet.

Bug: 276885705
Test: atest BluetoothInstrumentationTests
Change-Id: I58fc952ed6202d0586cf1e0fc7b7e2758f28e4aa