Minidroid: use vbmeta_system for verifying system
When both the avb, and the avb_keys= flags are specified then the latter
takes precedence, and an image is verified using the VBMeta struct
stored on the image itself, using the public keys located at the path on
the file system specified by the avb_keys= flag. On Minidroid, this way
of verification is incomplete:
- the VBMeta struct on system does not have an algortihm specified for
  the verification*,
- there are no keys stored at /avb.
Minidroid boots, but the image is not actually verified, and there are
errors logged during boot:

Error verifying vbmeta image: OK_NOT_SIGNED
Found unknown public key used to sign /system
Returning avb_handle for '/system' with status: VerificationError

So, for now, to fix the error and to simplify the setup, we'll only use
the vbmeta_system partition to verify system.

*see by running:
avbtool info_image --image out/target/product/sdv_core_cf/system.img

Bug: 281019451
Test: source build/envsetup.sh
Test: lunch aosp_cf_x86_64_minidroid-userdebug
Test: make
Test: launch_cvd
Test: # Check for AVB errors
Test: adb logcat | grep libfs_avb
Change-Id: If6e36c6dda97c8d33c1c392c40a4f720b28c51d1