Fix DirectAccessConnectionManager tests
1. Updating the state of connection to CONNECTED causes it to be added
to a map in connection manager(CM). CM also tracks devices that
disconnect from ADB to remove their entries from the map. If another
device connects to ADB, it will trigger the collect and disconnect the
connection whose state changed since the device might not have connected
to ADB yet.
Wait for device to be connected to Adb before updating the
state to CONNECTED.
2. When waiting for reservation to activate, the reservation might take
time to update in DirectAccessConnection. This might result in passing
older inactive connection to forwardingDaemonFactory. This reservation
will be used to retrieve connectionInfo which the old reservation will
not have resulting in a NoSuchElementException thrown while fetching
deviceAssociationName.
Force a state update after waitUntilActive() to pass latest reservation
to forwardingDaemonFactory

Bug: 279807103
Test: Updated
Change-Id: I4eb2b6b4c0c2a1438cc9d211a697ff613eb1b787